pyproject.toml
[project]
name = "agno_agent_app"
version = "0.1.0"
description = "AI Agent Web Application"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.7.7",
    "fastapi>=0.116.1",
    "google-genai>=1.28.0",
    "jinja2>=3.1.6",
    "markitdown[docx,pptx,xlsx]>=0.1.2",
    "matplotlib>=3.10.5",
    "mcp>=1.12.3",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "pywebview>=6.0",
    "sqlalchemy>=2.0.42",
    "tabulate>=0.9.0",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pyinstaller>=6.15.0",
]
